refactor: Rename hexagonal annotations sample package from `io.dragee` to `io.fixentropy` and update related approval resources and build configuration.
