+++ hexagonal-annotations/pom.xml
@@ -4,30 +4,30 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.dragee</groupId>
-        <artifactId>dragee-parent</artifactId>
+        <groupId>io.fixentropy</groupId>
+        <artifactId>fixentropy-parent</artifactId>
         <version>latest</version>
     </parent>
 
     <artifactId>hexagonal-annotations</artifactId>
     <packaging>jar</packaging>
     <name>hexagonal-annotations</name>
-    <url>https://dragee.io</url>
+    <url>https://fixentropy.io</url>
     <scm>
-        <connection>scm:git:git://github.com/dragee-io/dragee-java-processor.git</connection>
-        <developerConnection>scm:git:ssh://github.com:dragee-io/dragee-java-processor.git</developerConnection>
-        <url>https://github.com/dragee-io/dragee-java-processor/tree/main</url>
+        <connection>scm:git:git://github.com/fixentropy-io/fixentropy-java-processor.git</connection>
+        <developerConnection>scm:git:ssh://github.com:fixentropy-io/fixentropy-java-processor.git</developerConnection>
+        <url>https://github.com/fixentropy-io/fixentropy-java-processor/tree/main</url>
     </scm>
 
     <dependencies>
         <dependency>
-            <groupId>io.dragee</groupId>
+            <groupId>io.fixentropy</groupId>
             <artifactId>core-annotations</artifactId>
             <version>${project.parent.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>io.dragee</groupId>
+            <groupId>io.fixentropy</groupId>
             <artifactId>annotation-processor</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
