--- hexagonal-annotations/src/test/java/io/dragee/annotation/hexagonal/sample/AnInboundPort.java
@@ -1,7 +1,0 @@
-package io.dragee.annotation.hexagonal.sample;
-
-import io.dragee.annotation.hexagonal.Hexagonal;
-
-@Hexagonal.InboundPort
-public class AnInboundPort {
-}
