--- hexagonal-annotations/src/test/java/io/dragee/annotation/hexagonal/sample/AnOutboundPort.java
@@ -1,7 +1,0 @@
-package io.dragee.annotation.hexagonal.sample;
-
-import io.dragee.annotation.hexagonal.Hexagonal;
-
-@Hexagonal.OutboundPort
-public class AnOutboundPort {
-}
