+++ hexagonal-annotations/src/test/java/io/fixentropy/annotation/hexagonal/sample/ADomainEntity.java
@@ -1,0 +1,7 @@
+package io.fixentropy.annotation.hexagonal.sample;
+
+import io.fixentropy.annotation.hexagonal.Hexagonal;
+
+@Hexagonal.DomainEntity
+public class ADomainEntity {
+}
