+++ hexagonal-annotations/src/test/java/io/fixentropy/annotation/hexagonal/sample/AnOutboundAdapter.java
@@ -1,0 +1,7 @@
+package io.fixentropy.annotation.hexagonal.sample;
+
+import io.fixentropy.annotation.hexagonal.Hexagonal;
+
+@Hexagonal.OutboundAdapter
+public class AnOutboundAdapter {
+}
