+++ hexagonal-annotations/src/test/resources/approval/domain_entity.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.ADomainEntity","profile":"hexagonal/domain_entity","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.ADomainEntity","profile":"hexagonal/domain_entity","depends_on":{}}
