+++ hexagonal-annotations/src/test/resources/approval/domain_service.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.ADomainService","profile":"hexagonal/domain_service","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.ADomainService","profile":"hexagonal/domain_service","depends_on":{}}
