+++ hexagonal-annotations/src/test/resources/approval/inbound_adapter.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.AnInboundAdapter","profile":"hexagonal/inbound_adapter","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.AnInboundAdapter","profile":"hexagonal/inbound_adapter","depends_on":{}}
