+++ hexagonal-annotations/src/test/resources/approval/inbound_port.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.AnInboundPort","profile":"hexagonal/inbound_port","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.AnInboundPort","profile":"hexagonal/inbound_port","depends_on":{}}
