+++ hexagonal-annotations/src/test/resources/approval/outbound_adapter.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.AnOutboundAdapter","profile":"hexagonal/outbound_adapter","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.AnOutboundAdapter","profile":"hexagonal/outbound_adapter","depends_on":{}}
