+++ hexagonal-annotations/src/test/resources/approval/outbound_port.json
@@ -1,1 +1,1 @@
-{"name":"io.dragee.annotation.hexagonal.sample.AnOutboundPort","profile":"hexagonal/outbound_port","depends_on":{}}
+{"name":"io.fixentropy.annotation.hexagonal.sample.AnOutboundPort","profile":"hexagonal/outbound_port","depends_on":{}}
